test: update tests for new error message
See previous commit. Updated tests. As a result of splitting error
messages by "." I had a failing integration test due to the use of
"e.g." in an error message which I have substituted for "like" because
we don't have time approved to fix the issue right now. Refer to issue
5.