feat: implement pagination on My Renovations
Validate search has been put in its own file. Javascript specific to the
my renovation page is in `renovationPagination.js`. The id query param
in the url for the pagination footer is being set to null on the my
renovation page because I couldn't work out a better way to be able to
usue the same fragment for both the task cards and the my renovations
page. I think that won't matter in the end because the task card
pagination will change anyway.