feat: hook up team join fragment
Added extra model attributes required for the team join fragment.
Removed description as we don't actually support that and as far as I
know there is no AC specifying it. Fixed issue with the accept/decline
buttons on the fragment resolving to relative links. Added last visited
page session attribute for the view requests inbox.

There might be some duplication going on but I think it is ok and for
the most part unavoidable.