feat: initial hourly rate fix
Fix 'base case' of hourly rate 500 error bug

Took a look at the code and found that the server was storing excessively large values as infinity. A conditional checking if the value is equal to Float.Positive_Infinity was added (as any value less than infinity is still permitted by ACs).