test: adjust tests
Cleanup tests, unused dependencies and imports. Also added a check to
make sure we only have the invalid address error if a different address
error is not already there. This is because it only makes sense to tell
the user to enter a correct address if they have in fact provided a
valid one as defined by LocationService.validateLocation.