test: unit test for new service method
Wrote a unit test for the method that gets a map of contractors by team id

Currently does not pass due to the method being tested requiring contractors with non null ids, which may be impossible to facilitate. Will leave test commented out for now to prevent it breaking pipelines.