chore: remove unecessary code from cherry pick
Earlier I cherry-picked some code form U2013/plot-contractor-locations.
Some of this code is not related to this task and was causing sonarqube
to report low coverage even though this code is now covered in another
branch. I'm removing it here so that it will get merged back in later
when Ryan's branch is finished.